mm: Drop Google's picks and put do_wp_page back in line with stable
This causes TEEGRIS to crash out and die. The user-visible result is
that the lockscreen pin is claimed to be incorrect, making the tablet
unusable. The device may be unlocked once upon boot, if done quickly
enough, but not after relocking or sitting long enough after boot.

This is a squash of:
* Revert "UPSTREAM: mm: reuse only-pte-mapped KSM page in do_wp_page()"
  This reverts commit 1ce88842dc611003e1d15e2ed65457681f5b6207.
* Revert "BACKPORT: mm: do_wp_page() simplification"
  This reverts commit 1f444665ed850e46afc64974f251eeb37ad8feea.
* Revert "UPSTREAM: mm: fix misplaced unlock_page in do_wp_page()"
  This reverts commit c056f215dc999f2e12c453ad5e42951b32c6c071.

Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Change-Id: I83980aea81a0b667068acdf6ffc7d698ff1d51bf